Fix added sanity check to preprocessed compiler output

+++ pycuda/compiler.py
@@ -53,7 +53,11 @@ def preprocess_source(source, options, nvcc):
         from pycuda.driver import CompileError
         raise CompileError("nvcc preprocessing of %s failed" % source_path,
                            cmdline, stderr=stderr)
-
+    # sanity check
+    if len(stdout) < 0.5*len(source):
+        from pycuda.driver import CompileError
+        raise CompileError("nvcc preprocessing of %s failed with ridiculously small code output - likely unsupported compiler." % source_path, cmdline, stderr=stderr)
+        
     unlink(source_path)
 
     return stdout
